pyproject.toml
[project]
name = "cloud-data-engineering-technologies"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastavro>=1.12.1",
    "flask>=3.1.2",
    "google-cloud-bigquery>=3.39.0",
    "google-cloud-pubsub>=2.33.0",
    "google-cloud-storage>=3.7.0",
    "pyarrow>=22.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
